requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2